recommender.py
import pandas as pd
import numpy as np

from Model.pipeline import load_data, preprocess_data
from Model.collaborative_filtering import collaborative_filter, predict_ratings
from Model.clustering import cluster_providers, get_cluster_stats
from Model.pareto_optimization import pareto_optimize

class HealthcareRecommender:
    def __init__(self, data_dir='data', k=50, n_clusters=5):
        """
        Initialize the healthcare provider recommender system.
        
        Args:
            data_dir (str): Directory containing the data files
            k (int): Number of latent factors for SVD
            n_clusters (int): Number of provider clusters
        """
        # ...existing code...
    
    def recommend(self, patient_id, top_n=10):
        """
        Generate recommendations for a patient.
        
        Args:
            patient_id (int): ID of the patient
            top_n (int): Number of recommendations to return
            
        Returns:
            pd.DataFrame: Top recommended providers
        """
        # ...existing code...
